requirements.txt
pymupdf>=1.24.0
openai>=1.0.0
pydantic>=2.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pywebview>=5.0.0
